requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
coloredlogs==15.0.1
Django==4.1.3
django-filter==22.1
django-ratelimit==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.2
flatbuffers==23.5.26
httplib2==0.21.0
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
odict==1.9.0
onnxruntime==1.15.0
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
plumber==1.7
protobuf==4.23.3
psycopg2==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
requests==2.31.0
sqlparse==0.4.3
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing-extensions==4.6.3
urllib3==2.0.3
zope.component==5.0.1
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.6
zope.hookable==5.4
zope.interface==5.5.2
zope.lifecycleevent==4.4
zope.proxy==4.6.1
